feat: disabled deleteOrganisation code in the controller level

+++ src/main/java/uk/gov/hmcts/reform/professionalapi/controller/internal/OrganisationInternalController.java
@@ -60,6 +60,11 @@ public class OrganisationInternalController extends SuperController {
     @Value("${loggingComponentName}")
     protected String loggingComponentName;
 
+    @Value("${deleteOrganisationEnabled}")
+    private String deleteOrganisationEnabled;
+
+
+
 
     @ApiOperation(
             value = "Creates an Organisation",
@@ -387,23 +392,26 @@ public class OrganisationInternalController extends SuperController {
     )
     @ResponseStatus(value = HttpStatus.NO_CONTENT)
     @ResponseBody
-    @Secured("prd-admin1")
+    @Secured("prd-admin")
     public ResponseEntity<DeleteOrganisationResponse> deleteOrganisation(
             @Pattern(regexp = ORGANISATION_IDENTIFIER_FORMAT_REGEX, message = ORG_ID_VALIDATION_ERROR_MESSAGE)
             @PathVariable("orgId") @NotBlank String organisationIdentifier) {
 
         DeleteOrganisationResponse deleteOrganisationResponse = new DeleteOrganisationResponse();
-        //Received request to delete an organisation for internal user
-        Optional<Organisation> organisation = Optional.ofNullable(organisationService
+        if (deleteOrganisationEnabled.equals("true")) {
+            //Received request to delete an organisation for internal user
+            Optional<Organisation> organisation = Optional.ofNullable(organisationService
                     .getOrganisationByOrgIdentifier(organisationIdentifier));
 
-        if (!organisation.isPresent()) {
+            if (!organisation.isPresent()) {
 
-            throw new EmptyResultDataAccessException(1);
+                throw new EmptyResultDataAccessException(1);
+
+            }
+
+            deleteOrganisationResponse = organisationService.deleteOrganisation(organisation.get());
 
         }
-
-        deleteOrganisationResponse = organisationService.deleteOrganisation(organisation.get());
         return ResponseEntity
                 .status(deleteOrganisationResponse.getStatusCode())
                 .body(deleteOrganisationResponse);
